requirements.txt
# apt-get  install build-essential python-dev protobuf-compiler libprotobuf-dev
imposm.parser
elasticsearch

